Set explicit perms, owner and group on puppet.conf
When the umask on the server is 0077, the puppet.conf file has
permissions of 0600 and the file can't be read by the foreman-proxy or
puppet since the default owner is root.  By setting the perms to world
readable, we resolve these issues.

Since puppet also defaults to the file owner on the puppet master, it's
better to explicitly set the owner and group.