Make foreman integration optional
Instead of including foreman::params, the values are copied. That means
foreman::params no longer needs to be included.

The next use of foreman is in puppet::server::config where it sets up
reporting and ENC. This is wrapped in a new boolean $server_foreman. The
user will still need to take care of clearing $server_reports and
$server_external_nodes.

Fixes #154