workaround for very slow rspec run
It seems that a large `operatingsystem_support` section in metadata.json
takes a really long time to get parsed by rspec-puppet-facts.  Since
tests are only run on one instance of a supported OS for most of the
tests this passes an opts hash to `on_supported_os` which limits parsing
and speeds up the rspec run a lot.