travis fast finish and rspec options
Rename LIMIT_OS option to ONLY_OS and introduce the EXCLUDE_OS option
which allows for excluding platforms from the test run.

Fast finish is enabled in .travis.yml which was re-ordered to prefer
newer versions of Puppet and Ruby as well as only testing the latest
version of the larger distributions first to get faster feedback about
test failures.

All platforms are covered by running the tests in another matrix axis
which excludes the platforms that were previously tested.

Puppet 2.7 is only tested on CentOS 6 and Ubuntu 12 as the tests are
slow on that axis as well as Puppet 2.7 is end of life.