Allow to specify the listen ip address of the puppetserver
Such an attempt was done in the past (#200) but was very complex due to
the use of apache::vhost, which does mot allow mixing IP and non-IP
based vhosts.

This PR only takes care of the Puppetserver side and issues a warning if
users try to change the IP address with the puppetmaster implementation.