Add new server_use_legacy_auth_conf param
For puppetserver >= 2.2, /etc/puppetlabs/puppetserver/conf.d/auth.conf is
only used if 'use-legacy-auth-conf' is set to false in puppetserver.conf
https://docs.puppetlabs.com/puppetserver/latest/config_file_auth.html#aside-changes-to-authorization-in-puppet-server-220

Before this commit, use-legacy-auth-conf was not set in
puppetserver.conf and would default to 'true'.  By default, this commit
sets use-legacy-auth-conf to 'false'.

A new server_puppetserver_version parameter has also been added.
The parameter defaults to 2.3.1, but can be overriden if a user is
installing an older version (eg 2.1.1 which doesn't support
use-legacy-auth-conf)