puppetserver AIO uses a different vardir
puppetserver runs as the puppet user, the puppet agent normally runs as
the root user. When the two try to share the same vardir,
puppetserver (running as the puppet user) can't always read/write/create
directories if the puppet agent creates them as the root user. The
problem is described more thoroughly in
https://tickets.puppetlabs.com/browse/SERVER-357 .