Deprecating and replacing server_facts parameter
The $server_facts variable name is reserved by puppet-servers when trusted_server_facts is enabled.
Using this variable in a manifest causes compilation to fail with "Attempt to assign to a reserved variable name: server_facts"
This commit deprecates this variable and introduces the boolean $server_foreman_facts as its replacement.