460 - allow enabling crl when server_ca => false
change allows the CRL to be enabled when puppet_ca is disabled
additionally provided the ability to sync from the master of masters
the #{ssldir}/ca/ca_crl.pem contents to #{ssldir}/crl.pem. The one side
note is i chose to not notify the puppetserver service everytime the
crl.pem file is updated. This could cause undesired downtime of the
compile masters. Reference \#460 for addtiional information.