Do not manage Puppet Server web-routes.conf
Maintaining the web routes config file has led to a number of bugs
between versions of Puppet Server and will in the future with the
changes between 2.7.x and 5.0.x.

As there are no user-configurable options in the file and the
documentation states that it should not be modified, it is safer to use
the file shipped with the server package. This will ensure the
configuration matches the codebase, preventing startup errors from
config mismatches.