Always manage the puppet user
The use case is that puppet-certs deploys certificates with the owner
puppet and now contains a require => Class['puppet::server::install']
but if the User resource is managed then puppet can autorequire it and
there is less coupling between modules.

Previously we managed the user only if a git_repo was requested.