Fixes #25839 - Determine puppetserver version in server.pp
Currently we store some parameters in the installers answers file that
relate to the puppetserver version. This is because we determine them in
params.pp which means init.pp has them. The result is that when a user
upgrades, their answers need to be reset. Katello has a hook to always
do this, Foreman has nothing.

By defaulting to undef and moving the the determination into server.pp
we can avoid this whole problem.