Handle the pluginsync setting deprecation
In Puppet 4 the pluginsync setting was deprecated. At least in Puppet 5
it generates a deprecation warning and in Puppet 6 it was removed.
Because the default was already true, we only emit it if it's
non-default to avoid the deprecation warning. In Puppet 6 we can detect
an invalid setting so we hard fail.