Check CA path when creating the CA cert
When starting a new puppetserver as CA with existing CA files, the current code
tries to perform 'puppetserver ca setup' because the puppetserver's cert does
not exist yet. However, that command fails because the CA cert files exist and
can, correctly, not be overwritten.

This change checks for the CA cert, instead of the server's own cert.