Revert "Revert "Convert the systemd timer service to camptocamp/systemd""
This reverts commit 904e51b3e0b2fc8b2a3098ca981ad3e4339dd611.

Initially it was reverted since it led to a dependency cycle in the
Foreman installer. This was caused by a systemctl daemon-reload. Since
this was a bit of an edge case and not all users will hit this, it's
reinstated.

The cyclic dependency can also be mitigated by using camptocamp/systemd
3.x which no longer performs the exec. This only works on Puppet 6.1 or
newer though.