Support multiple envs_dir directories
This module adds support for the puppet environmentpath setting to be configured
using multiple directories. An example use-case is where r10k is in use to
deploy some but not all environments, to avoid unmanaged environments being
purged by r10k. Directories (whether one, or multiple) are passed as an
Array of Stdlib::Absolutepath values.

This is a breaking change, as a single string will no longer be accepted.

- Each listed environmentpath directory will be created and managed by puppet
- The default post-receive hook script uses the first listed directory in
  `envs_dir` to create initial environments to maintain backward compatibility.

Tests are included to ensure that all listed directories are created and
managed.

Fixes #708