Drop config_version and environment.conf
Support for dynamic and config environments was dropped in a61e010
however the `config_version` related parameters and the
`environment.conf` template used by these features were not removed
at the same time. This commit drops the unused code.

As parameters are being removed, this will be a breaking change and
require a major version bump.