puppet agent timer: depend on network/start it if offline
Two changes: With adding network-online.target as dependency we ensure
that the network is really available. That might not be the case for the
network.target. In case the system has no proper network manager,
network-online.target isn't available. For that situation we still
depend on network.target. Also we now have Wants= defined. That will
start the targets if the agent is started.

We had a similar discussion upstream:
https://github.com/puppetlabs/puppet/pull/8893