Avoid unless: kwarg on upgrade acceptance test
With the unless keyword argument on a describe block the content is
still parsed by Ruby. This means some functions are still executed. It
also ends up triggering the unsupported version exception on Archlinux
because the collection is unrecognized, even though the test really
isn't run. By using the real Ruby unless statement this is all avoided.